+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/client/HUD.kt
@@ -13,6 +13,7 @@ import net.ccbluex.liquidbounce.features.module.ModuleInfo
 import net.ccbluex.liquidbounce.features.module.modules.client.button.AbstractButtonRenderer
 import net.ccbluex.liquidbounce.features.module.modules.client.button.FLineButtonRenderer
 import net.ccbluex.liquidbounce.features.module.modules.client.button.RiseButtonRenderer
+import net.ccbluex.liquidbounce.features.module.modules.client.button.BetterButtonRenderer
 import net.ccbluex.liquidbounce.features.module.modules.client.button.RoundedButtonRenderer
 import net.ccbluex.liquidbounce.launch.data.legacyui.ClickGUIModule.*
 import net.ccbluex.liquidbounce.ui.client.hud.designer.GuiHudDesigner
@@ -59,7 +60,7 @@ object HUD : Module() {
     val arraylistYAxisAnimTypeValue = EaseUtils.getEnumEasingList("ArraylistYAxisAnimType")
     val arraylistYAxisAnimOrderValue = EaseUtils.getEnumEasingOrderList("ArraylistYAxisHotbarAnimOrder")
     val fontEpsilonValue = FloatValue("FontVectorEpsilon", 0.5f, 0f, 1.5f)
-    private val buttonValue = ListValue("Button", arrayOf("Better", "FLine", "Rise", "Vanilla"), "Better")
+    private val buttonValue = ListValue("Button", arrayOf("Better", "Rounded", "FLine", "Rise", "Vanilla"), "Rounded")
 
     private var lastFontEpsilon = 0f
 
@@ -162,7 +163,8 @@ object HUD : Module() {
 
     fun getButtonRenderer(button: GuiButton): AbstractButtonRenderer? {
         return when (buttonValue.get().lowercase()) {
-            "better" -> RoundedButtonRenderer(button)
+            "better" -> BetterButtonRenderer(button)
+            "rounded" -> RoundedButtonRenderer(button)
             "fline" -> FLineButtonRenderer(button)
             "rise" -> RiseButtonRenderer(button)
             else -> null // vanilla or unknown
